--- src/components/addmanual.scss
@@ -1,10 +1,0 @@
-.addmanualdrawer {
-  margin: 0 1em 1em 1em;
-  h3 {
-    text-align: center;
-  }
-  form {
-    display: grid;
-    row-gap: 1em;
-  }
-}
