+++ src/components/views/View1SSO.jsx
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React, { useState, useEffect } from "react";
 
 import { Button, TextField } from "@material-ui/core";
 
@@ -8,11 +8,18 @@ function View1SSO(props) {
 	const [sso, setSSO] = useState("");
 	const [inge, setInge] = useState(null);
 
+	useEffect(() => {
+		if (props.step === 0) {
+			inge ? props.handleNext(false) : props.handleNext(true);
+		}
+	}, [inge, props]);
+
 	const changeSSO = (value) => {
 		setSSO(value);
 		if (inge) {
-			handleDone(true, null);
 			setInge(null);
+			props.onDone(null);
+			props.handleNext(true);
 		}
 	};
 	const validateSSO = () => {
@@ -22,24 +29,22 @@ function View1SSO(props) {
 			.get()
 			.then((data) => {
 				if (data.empty) {
-					handleDone(true, null);
+					props.onDone(null);
+					props.handleNext(true);
 					alert("No existe el usuario");
 				} else {
 					data.forEach((user) => {
 						let us = { ...user.data(), uid: user.id };
 						console.log(us);
 						setInge(us);
-						handleDone(false, us);
+						props.onDone(us);
+						props.handleNext(false);
 					});
 				}
 			})
 			.catch((err) => {
 				console.log("Error: " + err);
 			});
-	};
-
-	const handleDone = (nD, user) => {
-		props.onDone(nD, user);
 	};
 
 	return (
@@ -69,7 +74,7 @@ function View1SSO(props) {
 					Validar
 				</Button>
 			)}
-			<b className="version">Version 1.2.0</b>
+			<b className="version">Version 1.2.2</b>
 		</>
 	);
 }
